+++ src/realm/object-store/sync/sync_manager.cpp
@@ -452,6 +452,41 @@ void SyncManager::remove_user(const std::string& user_id)
     }
 }
 
+void SyncManager::delete_user(const std::string& user_id)
+{
+    util::CheckedLockGuard lock(m_user_mutex);
+    // Avoid itterating over m_users twice by not calling `get_user_for_identity`.
+    auto it = std::find_if(m_users.begin(), m_users.end(), [&user_id](auto& user) {
+        return user->identity() == user_id;
+    });
+    auto user = it == m_users.end() ? nullptr : *it;
+
+    if (!user)
+        return;
+
+    // Deletion should happen immediately, not when we do the cleanup
+    // task on next launch.
+    m_users.erase(it);
+    user->detach_from_sync_manager();
+
+    if (m_current_user && m_current_user->identity() == user->identity())
+        m_current_user = nullptr;
+
+    util::CheckedLockGuard fs_lock(m_file_system_mutex);
+    if (!m_metadata_manager)
+        return;
+
+    auto users = m_metadata_manager->all_unmarked_users();
+    for (size_t i = 0; i < users.size(); i++) {
+        auto metadata = users.get(i);
+        if (user->identity() == metadata.identity()) {
+            m_file_manager->remove_user_realms(metadata.identity(), metadata.realm_file_paths());
+            metadata.remove();
+            break;
+        }
+    }
+}
+
 SyncManager::~SyncManager()
 {
     // Grab a vector of the current sessions under a lock so we can shut them down. We have to make a copy because